+++ Kernel/System/Console/Command/Help.pm
@@ -11,7 +11,10 @@ package Kernel::System::Console::Command::Help;
 use strict;
 use warnings;
 
-use base qw(Kernel::System::Console::BaseCommand);
+use base qw(
+    Kernel::System::Console::BaseCommand
+    Kernel::System::Console::Command::List
+);
 
 our @ObjectDependencies = (
     'Kernel::Config',
@@ -21,10 +24,10 @@ our @ObjectDependencies = (
 sub Configure {
     my ( $Self, %Param ) = @_;
 
-    $Self->Description('Displays help for a command.');
+    $Self->Description('Displays help for an existing command or search for commands.');
     $Self->AddArgument(
         Name        => 'command',
-        Description => "Print usage information for this command.",
+        Description => "Print usage information for this command (if command is available) or search for commands with similar names.",
         ValueRegex  => qr/[a-zA-Z0-9:_]+/,
         Required    => 1,
     );
@@ -33,10 +36,50 @@ sub Configure {
 sub Run {
     my ( $Self, %Param ) = @_;
 
-    my $Command = $Kernel::OM->Get( 'Kernel::System::Console::Command::' . $Self->GetArgument('command') );
-    $Command->ANSI( $Self->ANSI() );
-    print $Command->GetUsageHelp();
+    my $SearchCommand =  $Self->GetArgument('command');
+    my $CommandModule = "Kernel::System::Console::Command::$SearchCommand";
 
+    # Is it an existing command? Then show help for it.
+    if ($Kernel::OM->Get('Kernel::System::Main')->Require( $CommandModule, Silent => 1 )) {
+        my $Command = $Kernel::OM->Get( $CommandModule );
+        $Command->ANSI( $Self->ANSI() );
+        print $Command->GetUsageHelp();
+        return $Self->ExitCodeOk();
+    }
+
+    # Otherwise, search for commands with a similar name
+    $Self->Print("Searching for commands similar to '<yellow>$SearchCommand</yellow>'...\n");
+
+    my $PreviousCommandNameSpace = '';
+    my $UsageText;
+
+    COMMAND:
+    for my $Command ( $Self->ListAllCommands() ) {
+        my $CommandObject = $Kernel::OM->Get($Command);
+
+        if ( $Command !~ m{\Q$SearchCommand\E}smxi &&
+            $CommandObject->Description() !~ m{\Q$SearchCommand\E}smxi
+        ) {
+            next COMMAND;
+        }
+        my $CommandName   = $CommandObject->Name();
+
+        # Group by toplevel namespace
+        my ($CommandNamespace) = $CommandName =~ m/^([^:]+)::/smx;
+        $CommandNamespace //= '';
+        if ( $CommandNamespace ne $PreviousCommandNameSpace ) {
+            $UsageText .= "<yellow>$CommandNamespace</yellow>\n";
+            $PreviousCommandNameSpace = $CommandNamespace;
+        }
+        $UsageText .= sprintf( " <green>%-40s</green> - %s\n", $CommandName, $CommandObject->Description() );
+    }
+
+    if (!$UsageText) {
+        $Self->Print("<yellow>No commands found.</yellow>\n");
+        return $Self->ExitCodeOk();
+    }
+
+    $Self->Print($UsageText);
     return $Self->ExitCodeOk();
 }